Move node executable to node-tui
