+++ Cargo.toml
@@ -23,7 +23,8 @@ members = [
   "mempool",                      # Mempool interface and implementation.
   "merkletree",                   # Merkle tree implementation with merkle proofs.
   "mocks",                        # Mock implementations of our traits (used for testing)
-  "node",                         # Node executable.
+  "node",                         # Node lib.
+  "node-tui",                     # Node terminal binary.
   "p2p",                          # P2p communication interfaces and protocols.
   "p2p/backend-test-suite",       # P2p backend agnostic tests.
   "pos_accounting",               # PoS accounting and balances abstractions
