+++ node/Cargo.toml
@@ -33,5 +33,4 @@ fs4 = "0.6"
 [dev-dependencies]
 crypto = { path = "../crypto" }
 
-assert_cmd = "2"
 tempfile = "3.3"
